Refactor component and update stats display
