+++ client/src/components/Blogs/Blogs.tsx
@@ -18,7 +18,7 @@ const Blogs = ({ allBlogs, loading }: AllBlogProps) => {
           {allBlogs?.length > 0 ? (
             <div className=" grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-5 ">
               {allBlogs.map((blog) => (
-                <BlogCard key={blog._id} blog={blog}></BlogCard>
+                <BlogCard key={blog.slug} blog={blog}></BlogCard>
               ))}
             </div>
           ) : (
