+++ client/src/pages/DashBoardHome/DashBoardHome.tsx
@@ -1,28 +1,19 @@
-import { useQuery } from "@tanstack/react-query";
 import useAuth from "../../hooks/useAuth";
 import moment from "moment";
-import useAxiosSecure from "../../api/useAxiosSecure";
 import { Helmet } from "react-helmet";
-
-type DashboardData = {
-  total: number;
-  draft: number;
-  published: number;
-};
+import useAdmin from "../../hooks/useAdmin";
+import useUserStats from "../../hooks/useUserStats";
+import LoadingSpinner from "../../components/LoadingSpinner/LoadingSpinner";
 
 const DashBoardHome = () => {
   const { user } = useAuth();
   const currentDay = moment().format("dddd");
 
-  const axiosSecure = useAxiosSecure();
+  const { result } = useAdmin();
 
-  const { data = {} as DashboardData } = useQuery<DashboardData>({
-    queryKey: ["dashboard", user?.uid],
-    queryFn: async () => {
-      const response = await axiosSecure.get("/blogs/stats");
-      return response.data.data;
-    },
-  });
+  const { stats, isLoading } = useUserStats();
+
+  if (isLoading) return <LoadingSpinner />;
 
   return (
     <section className="dark:text-white">
@@ -32,7 +23,11 @@ const DashBoardHome = () => {
         </Helmet>
       </>
       <h1 className="text-xl font-semibold mb-2">
-        Welcome {user?.displayName}
+        Welcome{" "}
+        <span className="font-bold text-blue-600">
+          {result.isAdmin ? "Admin" : "User"}
+        </span>
+        <span> {user?.displayName}</span>
       </h1>
       <p className="text-sm font-semibold">Toaday is {currentDay}</p>
       <h1 className="mt-4 text-center font-bold text-lg">Current stats :</h1>
@@ -40,18 +35,19 @@ const DashBoardHome = () => {
         <div className="bg-[#E8E8EA] dark:bg-[#242535] p-3 rounded-sm max-w-52">
           <div className="flex justify-between ">
             <span>Total Post :</span>{" "}
-            <span>{data?.total ? data?.total : 0}</span>
+            <span>{stats?.total ? stats?.total : 0}</span>
           </div>
         </div>
         <div className="bg-[#E8E8EA] dark:bg-[#242535] p-3 rounded-sm max-w-52">
           <div className="flex justify-between ">
-            <span>Pending :</span> <span>{data?.draft ? data?.draft : 0}</span>
+            <span>Pending :</span>{" "}
+            <span>{stats?.draft ? stats?.draft : 0}</span>
           </div>
         </div>
         <div className="bg-[#E8E8EA] dark:bg-[#242535] p-3 rounded-sm max-w-52">
           <div className="flex justify-between ">
             <span>Published :</span>{" "}
-            <span>{data?.published ? data?.published : 0}</span>
+            <span>{stats?.published ? stats?.published : 0}</span>
           </div>
         </div>
       </div>
